Add a configure_requires dependency on ExtUtils::F77
This solves the problem of CPAN tools running Build.PL, finding that
ExtUtils::F77 is required, installing it and then not running
Build.PL again.

The downside is that I think it means that every user will be
forced to install ExtUtils::F77 even if they are using the C
library because the installers scan the META.yml files for
configure dependencies before even attempting to run Build.PL.

Currently, the META.yml is filled in based on the initial
configuration of Build.PL when I make the distribution. This
forces ExtUtils::F77 on CPAN as a dependency but if I test
against the C library before making the distribution that
dependency will disappear.